# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NanoEdgeRT API",
    "description": "NanoEdgeRT is a dynamic service management platform that allows you to create, manage, and run JavaScript-based microservices with JWT authentication, OpenAPI documentation, and real-time service orchestration.",
    "version": "2.6.0",
    "contact": { "name": "NanoEdgeRT", "url": "https://github.com/LemonHX/NanoEdgeRT" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [{ "url": "http://127.0.0.1:8000", "description": "Local development server" }],
  "paths": {
...
    "/admin-api/v2/services": {
      "get": {
        "summary": "Get all services",
        "description": "Retrieve a list of all registered services",
        "operationId": "getAllServices",
        "tags": ["Admin - Services"],
        "security": [{ "jwtAuth": [] }],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Service" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new service",
        "description": "Create a new service with the specified configuration",
        "operationId": "createService",
        "tags": ["Admin - Services"],
        "security": [{ "jwtAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateServiceRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" }, "name": { "type": "string" } }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/admin-api/v2/services/{name}": {
      "get": {
        "summary": "Get service by name",
        "description": "Retrieve a specific service by its name",
        "operationId": "getService",
        "tags": ["Admin - Services"],
        "security": [{ "jwtAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Name of the service",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Service" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "put": {
        "summary": "Update service",
        "description": "Update an existing service configuration",
        "operationId": "updateService",
        "tags": ["Admin - Services"],
        "security": [{ "jwtAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Name of the service",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateServiceRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "message": { "type": "string" } } }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete service",
        "description": "Delete a service by its name",
        "operationId": "deleteService",
        "tags": ["Admin - Services"],
        "security": [{ "jwtAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Name of the service",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "message": { "type": "string" } } }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
    "/admin-api/v2/config/{key}": {
      "get": {
        "summary": "Get configuration value",
        "description": "Retrieve a specific configuration value by key",
        "operationId": "getConfig",
        "tags": ["Admin - Configuration"],
        "security": [{ "jwtAuth": [] }],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "Configuration key",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": { "type": "string" },
                    "value": {
                      "oneOf": [
                        { "type": "string" },
                        { "type": "number" },
                        { "type": "boolean" },
                        { "type": "object" }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Configuration key not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "put": {
        "summary": "Update configuration value",
        "description": "Update a specific configuration value by key",
        "operationId": "updateConfig",
        "tags": ["Admin - Configuration"],
        "security": [{ "jwtAuth": [] }],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "Configuration key",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "oneOf": [
                      { "type": "string" },
                      { "type": "number" },
                      { "type": "boolean" },
                      { "type": "object" }
                    ]
                  }
                },
                "required": ["value"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "message": { "type": "string" } } }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication. Include the token in the Authorization header as 'Bearer <token>'."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" },
          "message": { "type": "string", "description": "Detailed error message" },
          "details": { "type": "string", "description": "Additional error details" }
        },
        "required": ["error"]
      },
...
      "ServicePermissions": {
        "type": "object",
        "properties": {
          "read": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Read permissions"
          },
          "write": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Write permissions"
          },
          "env": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Environment variable permissions"
          },
          "run": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Execution permissions"
          }
        },
        "required": ["read", "write", "env", "run"]
      },
      "Service": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Service name" },
          "code": { "type": "string", "description": "JavaScript code for the service" },
          "enabled": { "type": "boolean", "description": "Whether the service is enabled" },
          "jwt_check": {
            "type": "boolean",
            "description": "Whether JWT authentication is required"
          },
          "permissions": { "$ref": "#/components/schemas/ServicePermissions" },
          "schema": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI schema JSON string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Service creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Service last update timestamp"
          }
        },
        "required": ["name", "code", "enabled", "jwt_check", "permissions"]
      },
      "CreateServiceRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Service name" },
          "code": { "type": "string", "description": "JavaScript code for the service" },
          "enabled": {
            "type": "boolean",
            "default": true,
            "description": "Whether the service is enabled"
          },
          "jwt_check": {
            "type": "boolean",
            "default": false,
            "description": "Whether JWT authentication is required"
          },
          "permissions": { "$ref": "#/components/schemas/ServicePermissions" },
          "schema": { "type": "string", "description": "OpenAPI schema JSON string" }
        },
        "required": ["name", "code"]
      },
      "UpdateServiceRequest": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "description": "JavaScript code for the service" },
          "enabled": { "type": "boolean", "description": "Whether the service is enabled" },
          "jwt_check": {
            "type": "boolean",
            "description": "Whether JWT authentication is required"
          },
          "permissions": { "$ref": "#/components/schemas/ServicePermissions" },
          "schema": { "type": "string", "description": "OpenAPI schema JSON string" }
        }
      },
...
    }
  },
...
}